[Refactor] supporting generator
