+++ rl4co/envs/routing/atsp/render.py
@@ -1,0 +1,50 @@
+import torch
+import numpy as np
+import matplotlib.pyplot as plt
+
+from rl4co.utils.ops import gather_by_index
+from rl4co.utils.pylogger import get_pylogger
+
+log = get_pylogger(__name__)
+
+
+def render(td, actions=None, ax=None):
+    if ax is None:
+        # Create a plot of the nodes
+        _, ax = plt.subplots()
+
+    td = td.detach().cpu()
+
+    if actions is None:
+        actions = td.get("action", None)
+
+    # If batch_size greater than 0 , we need to select the first batch element
+    if td.batch_size != torch.Size([]):
+        td = td[0]
+        actions = actions[0]
+
+    locs = td["locs"]
+
+    # Gather locs in order of action if available
+    if actions is None:
+        log.warning("No action in TensorDict, rendering unsorted locs")
+    else:
+        actions = actions.detach().cpu()
+        locs = gather_by_index(locs, actions, dim=0)
+
+    # Cat the first node to the end to complete the tour
+    locs = torch.cat((locs, locs[0:1]))
+    x, y = locs[:, 0], locs[:, 1]
+
+    # Plot the visited nodes
+    ax.scatter(x, y, color="tab:blue")
+
+    # Add arrows between visited nodes as a quiver plot
+    dx, dy = np.diff(x), np.diff(y)
+    ax.quiver(
+        x[:-1], y[:-1], dx, dy, scale_units="xy", angles="xy", scale=1, color="k"
+    )
+
+    # Setup limits and show
+    ax.set_xlim(-0.05, 1.05)
+    ax.set_ylim(-0.05, 1.05)
