+++ rl4co/envs/routing/cvrp/env.py
@@ -15,26 +15,10 @@ from rl4co.envs.common.base import RL4COEnvBase
 from rl4co.utils.ops import gather_by_index, get_tour_length
 from rl4co.utils.pylogger import get_pylogger
 
+from .generator import CVRPGenerator
+from .render import render
+
 log = get_pylogger(__name__)
-
-
-# From Kool et al. 2019, Hottung et al. 2022, Kim et al. 2023
-CAPACITIES = {
-    10: 20.0,
-    15: 25.0,
-    20: 30.0,
-    30: 33.0,
-    40: 37.0,
-    50: 40.0,
-    60: 43.0,
-    75: 45.0,
-    100: 50.0,
-    125: 55.0,
-    150: 60.0,
-    200: 70.0,
-    500: 100.0,
-    1000: 150.0,
-}
 
 
 class CVRPEnv(RL4COEnvBase):
@@ -58,29 +42,15 @@ class CVRPEnv(RL4COEnvBase):
 
     def __init__(
         self,
-        num_loc: int = 20,
-        min_loc: float = 0,
-        max_loc: float = 1,
-        min_demand: float = 1,
-        max_demand: float = 10,
-        vehicle_capacity: float = 1.0,
-        capacity: float = None,
-        td_params: TensorDict = None,
+        generator: CVRPGenerator = None,
+        generator_params: dict = {},
         **kwargs,
     ):
         super().__init__(**kwargs)
-        self.num_loc = num_loc
-        self.min_loc = min_loc
-        self.max_loc = max_loc
-        self.min_demand = min_demand
-        self.max_demand = max_demand
-        self.capacity = CAPACITIES.get(num_loc, None) if capacity is None else capacity
-        if self.capacity is None:
-            raise ValueError(
-                f"Capacity for {num_loc} locations is not defined. Please provide a capacity manually."
-            )
-        self.vehicle_capacity = vehicle_capacity
-        self._make_spec(td_params)
+        if generator is None:
+            generator = CVRPGenerator(**generator_params)
+        self.generator = generator
+        self._make_spec(self.generator)
 
     def _step(self, td: TensorDict) -> TensorDict:
         current_node = td["action"][:, None]  # Add dimension for step
@@ -121,13 +91,7 @@ class CVRPEnv(RL4COEnvBase):
         td: Optional[TensorDict] = None,
         batch_size: Optional[list] = None,
     ) -> TensorDict:
-        if batch_size is None:
-            batch_size = self.batch_size if td is None else td["locs"].shape[:-2]
-        if td is None or td.is_empty():
-            td = self.generate_data(batch_size=batch_size)
-        batch_size = [batch_size] if isinstance(batch_size, int) else batch_size
-
-        self.to(td.device)
+        device = td.device
 
         # Create reset TensorDict
         td_reset = TensorDict(
@@ -135,16 +99,16 @@ class CVRPEnv(RL4COEnvBase):
                 "locs": torch.cat((td["depot"][:, None, :], td["locs"]), -2),
                 "demand": td["demand"],
                 "current_node": torch.zeros(
-                    *batch_size, 1, dtype=torch.long, device=self.device
+                    *batch_size, 1, dtype=torch.long, device=device
                 ),
-                "used_capacity": torch.zeros((*batch_size, 1), device=self.device),
+                "used_capacity": torch.zeros((*batch_size, 1), device=device),
                 "vehicle_capacity": torch.full(
-                    (*batch_size, 1), self.vehicle_capacity, device=self.device
+                    (*batch_size, 1), self.generator.vehicle_capacity, device=device
                 ),
                 "visited": torch.zeros(
                     (*batch_size, 1, td["locs"].shape[-2] + 1),
                     dtype=torch.uint8,
-                    device=self.device,
+                    device=device,
                 ),
             },
             batch_size=batch_size,
@@ -166,11 +130,7 @@ class CVRPEnv(RL4COEnvBase):
         mask_depot = (td["current_node"] == 0) & ((mask_loc == 0).int().sum(-1) > 0)
         return ~torch.cat((mask_depot[..., None], mask_loc), -1).squeeze(-2)
 
-    def get_reward(self, td: TensorDict, actions: TensorDict) -> TensorDict:
-        # Check that the solution is valid
-        if self.check_solution:
-            self.check_solution_validity(td, actions)
-
+    def _get_reward(self, td: TensorDict, actions: TensorDict) -> TensorDict:
         # Gather dataset in order of tour
         batch_size = td["locs"].shape[0]
         depot = td["locs"][..., 0:1, :]
@@ -215,48 +175,6 @@ class CVRPEnv(RL4COEnvBase):
                 used_cap <= td["vehicle_capacity"] + 1e-5
             ).all(), "Used more than capacity"
 
-    def generate_data(self, batch_size) -> TensorDict:
-        # Batch size input check
-        batch_size = [batch_size] if isinstance(batch_size, int) else batch_size
-
-        # Initialize the locations (including the depot which is always the first node)
-        locs_with_depot = (
-            torch.FloatTensor(*batch_size, self.num_loc + 1, 2)
-            .uniform_(self.min_loc, self.max_loc)
-            .to(self.device)
-        )
-
-        # Initialize the demand for nodes except the depot
-        # Demand sampling Following Kool et al. (2019)
-        # Generates a slightly different distribution than using torch.randint
-        demand = (
-            (
-                torch.FloatTensor(*batch_size, self.num_loc)
-                .uniform_(self.min_demand - 1, self.max_demand - 1)
-                .int()
-                + 1
-            )
-            .float()
-            .to(self.device)
-        )
-
-        # Support for heterogeneous capacity if provided
-        if not isinstance(self.capacity, torch.Tensor):
-            capacity = torch.full((*batch_size,), self.capacity, device=self.device)
-        else:
-            capacity = self.capacity
-
-        return TensorDict(
-            {
-                "locs": locs_with_depot[..., 1:, :],
-                "depot": locs_with_depot[..., 0, :],
-                "demand": demand / self.capacity,
-                "capacity": capacity,
-            },
-            batch_size=batch_size,
-            device=self.device,
-        )
-
     @staticmethod
     def load_data(fpath, batch_size=[]):
         """Dataset loading from file
@@ -266,13 +184,12 @@ class CVRPEnv(RL4COEnvBase):
         td_load.set("demand", td_load["demand"] / td_load["capacity"][:, None])
         return td_load
 
-    def _make_spec(self, td_params: TensorDict):
-        """Make the observation and action specs from the parameters."""
+    def _make_spec(self, generator: CVRPGenerator):
         self.observation_spec = CompositeSpec(
             locs=BoundedTensorSpec(
-                low=self.min_loc,
-                high=self.max_loc,
-                shape=(self.num_loc + 1, 2),
+                low=generator.min_loc,
+                high=generator.max_loc,
+                shape=(generator.num_loc + 1, 2),
                 dtype=torch.float32,
             ),
             current_node=UnboundedDiscreteTensorSpec(
@@ -280,13 +197,13 @@ class CVRPEnv(RL4COEnvBase):
                 dtype=torch.int64,
             ),
             demand=BoundedTensorSpec(
-                low=-self.capacity,
-                high=self.max_demand,
-                shape=(self.num_loc, 1),  # demand is only for customers
+                low=-generator.capacity,
+                high=generator.max_demand,
+                shape=(generator.num_loc + 1, 1),
                 dtype=torch.float32,
             ),
             action_mask=UnboundedDiscreteTensorSpec(
-                shape=(self.num_loc + 1, 1),
+                shape=(generator.num_loc + 1, 1),
                 dtype=torch.bool,
             ),
             shape=(),
@@ -295,143 +212,11 @@ class CVRPEnv(RL4COEnvBase):
             shape=(1,),
             dtype=torch.int64,
             low=0,
-            high=self.num_loc + 1,
+            high=generator.num_loc + 1,
         )
         self.reward_spec = UnboundedContinuousTensorSpec(shape=(1,))
         self.done_spec = UnboundedDiscreteTensorSpec(shape=(1,), dtype=torch.bool)
 
     @staticmethod
-    def render(
-        td: TensorDict,
-        actions=None,
-        ax=None,
-        scale_xy: bool = True,
-    ):
-        import matplotlib.pyplot as plt
-        import numpy as np
-
-        from matplotlib import cm, colormaps
-
-        num_routine = (actions == 0).sum().item() + 2
-        base = colormaps["nipy_spectral"]
-        color_list = base(np.linspace(0, 1, num_routine))
-        cmap_name = base.name + str(num_routine)
-        out = base.from_list(cmap_name, color_list, num_routine)
-
-        if ax is None:
-            # Create a plot of the nodes
-            _, ax = plt.subplots()
-
-        td = td.detach().cpu()
-
-        if actions is None:
-            actions = td.get("action", None)
-
-        # if batch_size greater than 0 , we need to select the first batch element
-        if td.batch_size != torch.Size([]):
-            td = td[0]
-            actions = actions[0]
-
-        locs = td["locs"]
-        scale_demand = CAPACITIES.get(td["locs"].size(-2) - 1, 1)
-        demands = td["demand"] * scale_demand
-
-        # add the depot at the first action and the end action
-        actions = torch.cat([torch.tensor([0]), actions, torch.tensor([0])])
-
-        # gather locs in order of action if available
-        if actions is None:
-            log.warning("No action in TensorDict, rendering unsorted locs")
-        else:
-            locs = locs
-
-        # Cat the first node to the end to complete the tour
-        x, y = locs[:, 0], locs[:, 1]
-
-        # plot depot
-        ax.scatter(
-            locs[0, 0],
-            locs[0, 1],
-            edgecolors=cm.Set2(2),
-            facecolors="none",
-            s=100,
-            linewidths=2,
-            marker="s",
-            alpha=1,
-        )
-
-        # plot visited nodes
-        ax.scatter(
-            x[1:],
-            y[1:],
-            edgecolors=cm.Set2(0),
-            facecolors="none",
-            s=50,
-            linewidths=2,
-            marker="o",
-            alpha=1,
-        )
-
-        # plot demand bars
-        for node_idx in range(1, len(locs)):
-            ax.add_patch(
-                plt.Rectangle(
-                    (locs[node_idx, 0] - 0.005, locs[node_idx, 1] + 0.015),
-                    0.01,
-                    demands[node_idx - 1] / (scale_demand * 10),
-                    edgecolor=cm.Set2(0),
-                    facecolor=cm.Set2(0),
-                    fill=True,
-                )
-            )
-
-        # text demand
-        for node_idx in range(1, len(locs)):
-            ax.text(
-                locs[node_idx, 0],
-                locs[node_idx, 1] - 0.025,
-                f"{demands[node_idx-1].item():.2f}",
-                horizontalalignment="center",
-                verticalalignment="top",
-                fontsize=10,
-                color=cm.Set2(0),
-            )
-
-        # text depot
-        ax.text(
-            locs[0, 0],
-            locs[0, 1] - 0.025,
-            "Depot",
-            horizontalalignment="center",
-            verticalalignment="top",
-            fontsize=10,
-            color=cm.Set2(2),
-        )
-
-        # plot actions
-        color_idx = 0
-        for action_idx in range(len(actions) - 1):
-            if actions[action_idx] == 0:
-                color_idx += 1
-            from_loc = locs[actions[action_idx]]
-            to_loc = locs[actions[action_idx + 1]]
-            ax.plot(
-                [from_loc[0], to_loc[0]],
-                [from_loc[1], to_loc[1]],
-                color=out(color_idx),
-                lw=1,
-            )
-            ax.annotate(
-                "",
-                xy=(to_loc[0], to_loc[1]),
-                xytext=(from_loc[0], from_loc[1]),
-                arrowprops=dict(arrowstyle="-|>", color=out(color_idx)),
-                size=15,
-                annotation_clip=False,
-            )
-
-        # Setup limits and show
-        if scale_xy:
-            ax.set_xlim(-0.05, 1.05)
-            ax.set_ylim(-0.05, 1.05)
-        plt.show()
+    def render(td: TensorDict, actions: torch.Tensor=None, ax = None):
+        return render(td, actions, ax)
