+++ rl4co/envs/routing/mpdp/env.py
@@ -14,6 +14,9 @@ from rl4co.envs.common.base import RL4COEnvBase
 from rl4co.utils.ops import gather_by_index
 from rl4co.utils.pylogger import get_pylogger
 
+from .generator import MPDPGenerator
+from .render import render
+
 log = get_pylogger(__name__)
 
 
@@ -40,25 +43,17 @@ class MPDPEnv(RL4COEnvBase):
 
     def __init__(
         self,
-        num_loc: int = 20,
-        min_loc: float = 0,
-        max_loc: float = 1,
-        min_num_agents: int = 2,
-        max_num_agents: int = 10,
+        generator: MPDPGenerator = None,
+        generator_params: dict = {},
         objective: str = "minmax",
-        check_solution: bool = False,
-        td_params: TensorDict = None,
         **kwargs,
     ):
         super().__init__(**kwargs)
-        self.num_loc = num_loc
-        self.min_loc = min_loc
-        self.max_loc = max_loc
-        self.min_num_agents = min_num_agents
-        self.max_num_agents = max_num_agents
+        if generator is None:
+            generator = MPDPGenerator(**generator_params)
+        self.generator = generator
         self.objective = objective
-        self.check_solution = check_solution
-        self._make_spec(td_params)
+        self._make_spec(self.generator)
 
     def _step(self, td: TensorDict) -> TensorDict:
         selected = td["action"][:, None]  # Add dimension for step
@@ -137,13 +132,7 @@ class MPDPEnv(RL4COEnvBase):
         batch_size: Optional[list] = None,
         agent_num: Optional[int] = None,  # NOTE hardcoded from ET
     ) -> TensorDict:
-        if batch_size is None:
-            batch_size = self.batch_size if td is None else td["locs"].shape[:-2]
-
-        if td is None or td.is_empty():
-            td = self.generate_data(batch_size=batch_size)
-
-        self.to(td.device)
+        device = td.device
 
         # NOTE: this is a hack to get the agent_num
         # agent_num = td["agent_num"][0].item() if agent_num is None else agent_num
@@ -157,7 +146,7 @@ class MPDPEnv(RL4COEnvBase):
 
         # Distance from all nodes between each other
         distance = torch.cdist(whole_instance, whole_instance, p=2)
-        index = torch.arange(left_request, 2 * left_request, device=depot.device)[
+        index = torch.arange(left_request, 2 * left_request, device=device)[
             None, :, None
         ]
         index = index.repeat(distance.shape[0], 1, 1)
@@ -195,10 +184,10 @@ class MPDPEnv(RL4COEnvBase):
                     1,
                     n_loc // 2 + agent_num + 1,
                     dtype=torch.uint8,
-                    device=loc.device,
+                    device=device,
                 ),
                 torch.zeros(
-                    batch_size, 1, n_loc // 2, dtype=torch.uint8, device=loc.device
+                    batch_size, 1, n_loc // 2, dtype=torch.uint8, device=device
                 ),
             ],
             dim=-1,
@@ -213,25 +202,25 @@ class MPDPEnv(RL4COEnvBase):
                     1,
                     n_loc + agent_num + 1,
                     dtype=torch.uint8,
-                    device=loc.device,
+                    device=device,
                 ),
-                "lengths": torch.zeros(batch_size, agent_num, device=loc.device),
-                "longest_lengths": torch.zeros(batch_size, agent_num, device=loc.device),
+                "lengths": torch.zeros(batch_size, agent_num, device=device),
+                "longest_lengths": torch.zeros(batch_size, agent_num, device=device),
                 "cur_coord": td["depot"]
                 if len(td["depot"].shape) == 2
                 else td["depot"].squeeze(1),
                 "i": torch.zeros(
-                    batch_size, dtype=torch.int64, device=loc.device
+                    batch_size, dtype=torch.int64, device=device
                 ),  # Vector with length num_steps
                 "to_delivery": to_delivery,
                 "count_depot": torch.zeros(
-                    batch_size, 1, dtype=torch.int64, device=loc.device
+                    batch_size, 1, dtype=torch.int64, device=device
                 ),
                 "agent_idx": torch.ones(
-                    batch_size, 1, dtype=torch.long, device=loc.device
+                    batch_size, 1, dtype=torch.long, device=device
                 ),
                 "left_request": left_request
-                * torch.ones(batch_size, 1, dtype=torch.long, device=loc.device),
+                * torch.ones(batch_size, 1, dtype=torch.long, device=device),
                 "remain_pickup_max_distance": remain_pickup_max_distance,
                 "remain_delivery_max_distance": remain_delivery_max_distance,
                 "depot_distance": depot_distance,
@@ -296,11 +285,7 @@ class MPDPEnv(RL4COEnvBase):
         action_mask = mask_loc == 0  # action_mask gets feasible actions
         return action_mask
 
-    def get_reward(self, td: TensorDict, actions: TensorDict) -> TensorDict:
-        # Check that the solution is valid
-        if self.check_solution:
-            self.check_solution_validity(td, actions)
-
+    def _get_reward(self, td: TensorDict, actions: TensorDict) -> TensorDict:
         # Calculate the reward (negative tour length)
         if self.objective == "minmax":
             return -td["lengths"].max(dim=-1, keepdim=True)[0].squeeze(-1)
@@ -313,32 +298,13 @@ class MPDPEnv(RL4COEnvBase):
     def check_solution_validity(td: TensorDict, actions: torch.Tensor):
         assert True, "Not implemented"
 
-    def generate_data(self, batch_size) -> TensorDict:
-        # Batch size input check
-        batch_size = [batch_size] if isinstance(batch_size, int) else batch_size
-
-        # Initialize the locations (including the depot which is always the first node)
-        locs_with_depot = (
-            torch.FloatTensor(*batch_size, self.num_loc + 1, 2)
-            .uniform_(self.min_loc, self.max_loc)
-            .to(self.device)
-        )
-
-        return TensorDict(
-            {
-                "locs": locs_with_depot[..., 1:, :],
-                "depot": locs_with_depot[..., 0, :],
-            },
-            batch_size=batch_size,
-        )
-
-    def _make_spec(self, td_params: TensorDict):
+    def _make_spec(self, generator: MPDPGenerator):
         """Make the observation and action specs from the parameters."""
         max_nodes = self.num_loc + self.max_num_agents + 1
         self.observation_spec = CompositeSpec(
             locs=BoundedTensorSpec(
-                low=self.min_loc,
-                high=self.max_loc,
+                low=generator.min_loc,
+                high=generator.max_loc,
                 shape=(max_nodes, 2),
                 dtype=torch.float32,
             ),
@@ -355,16 +321,16 @@ class MPDPEnv(RL4COEnvBase):
                 dtype=torch.bool,
             ),
             lengths=UnboundedContinuousTensorSpec(
-                shape=(self.max_num_agents,),
+                shape=(generator.max_num_agents,),
                 dtype=torch.float32,
             ),
             longest_lengths=UnboundedContinuousTensorSpec(
-                shape=(self.max_num_agents,),
+                shape=(generator.max_num_agents,),
                 dtype=torch.float32,
             ),
             cur_coord=BoundedTensorSpec(
-                low=self.min_loc,
-                high=self.max_loc,
+                low=generator.min_loc,
+                high=generator.max_loc,
                 shape=(2,),
                 dtype=torch.float32,
             ),
@@ -424,105 +390,5 @@ class MPDPEnv(RL4COEnvBase):
         self.done_spec = UnboundedDiscreteTensorSpec(shape=(1,), dtype=torch.bool)
 
     @staticmethod
-    def render(td: TensorDict, actions=None, ax=None):
-        # TODO: color switch with new agents; add pickup and delivery nodes as in `PDPEnv.render`
-
-        import matplotlib.pyplot as plt
-        import numpy as np
-
-        from matplotlib import cm, colormaps
-
-        num_routine = (actions == 0).sum().item() + 2
-        base = colormaps["nipy_spectral"]
-        color_list = base(np.linspace(0, 1, num_routine))
-        cmap_name = base.name + str(num_routine)
-        out = base.from_list(cmap_name, color_list, num_routine)
-
-        if ax is None:
-            # Create a plot of the nodes
-            _, ax = plt.subplots()
-
-        td = td.detach().cpu()
-
-        if actions is None:
-            actions = td.get("action", None)
-
-        # if batch_size greater than 0 , we need to select the first batch element
-        if td.batch_size != torch.Size([]):
-            td = td[0]
-            actions = actions[0]
-
-        locs = td["locs"]
-
-        # add the depot at the first action and the end action
-        actions = torch.cat([torch.tensor([0]), actions, torch.tensor([0])])
-
-        # gather locs in order of action if available
-        if actions is None:
-            log.warning("No action in TensorDict, rendering unsorted locs")
-        else:
-            locs = locs
-
-        # Cat the first node to the end to complete the tour
-        x, y = locs[:, 0], locs[:, 1]
-
-        # plot depot
-        ax.scatter(
-            locs[0, 0],
-            locs[0, 1],
-            edgecolors=cm.Set2(2),
-            facecolors="none",
-            s=100,
-            linewidths=2,
-            marker="s",
-            alpha=1,
-        )
-
-        # plot visited nodes
-        ax.scatter(
-            x[1:],
-            y[1:],
-            edgecolors=cm.Set2(0),
-            facecolors="none",
-            s=50,
-            linewidths=2,
-            marker="o",
-            alpha=1,
-        )
-
-        # text depot
-        ax.text(
-            locs[0, 0],
-            locs[0, 1] - 0.025,
-            "Depot",
-            horizontalalignment="center",
-            verticalalignment="top",
-            fontsize=10,
-            color=cm.Set2(2),
-        )
-
-        # plot actions
-        color_idx = 0
-        for action_idx in range(len(actions) - 1):
-            if actions[action_idx] == 0:
-                color_idx += 1
-            from_loc = locs[actions[action_idx]]
-            to_loc = locs[actions[action_idx + 1]]
-            ax.plot(
-                [from_loc[0], to_loc[0]],
-                [from_loc[1], to_loc[1]],
-                color=out(color_idx),
-                lw=1,
-            )
-            ax.annotate(
-                "",
-                xy=(to_loc[0], to_loc[1]),
-                xytext=(from_loc[0], from_loc[1]),
-                arrowprops=dict(arrowstyle="-|>", color=out(color_idx)),
-                size=15,
-                annotation_clip=False,
-            )
-
-        # Setup limits and show
-        ax.set_xlim(-0.05, 1.05)
-        ax.set_ylim(-0.05, 1.05)
+    def render(td: TensorDict, actions: torch.Tensor=None, ax = None):
+        return render(td, actions, ax)
