+++ rl4co/envs/routing/spctsp/env.py
@@ -1,6 +1,6 @@
 from rl4co.utils.pylogger import get_pylogger
 
-from .pctsp import PCTSPEnv
+from ..pctsp.env import PCTSPEnv
 
 log = get_pylogger(__name__)
