+++ rl4co/envs/routing/svrp/env.py
@@ -15,6 +15,9 @@ from rl4co.envs.common.base import RL4COEnvBase
 from rl4co.utils.ops import gather_by_index, get_distance
 from rl4co.utils.pylogger import get_pylogger
 
+from .generator import SVRPGenerator
+from .render import render
+
 log = get_pylogger(__name__)
 
 
@@ -41,32 +44,24 @@ class SVRPEnv(RL4COEnvBase):
 
     def __init__(
         self,
-        num_loc: int = 20,
-        min_loc: float = 0,
-        max_loc: float = 1,
-        min_skill: float = 1,
-        max_skill: float = 10,
-        tech_costs: list = [1, 2, 3],
-        td_params: TensorDict = None,
+        generator: SVRPGenerator = None,
+        generator_params: dict = {},
         **kwargs,
     ):
         super().__init__(**kwargs)
-        self.num_loc = num_loc
-        self.min_loc = min_loc
-        self.max_loc = max_loc
-        self.min_skill = min_skill
-        self.max_skill = max_skill
-        self.tech_costs = tech_costs
-        self.num_tech = len(tech_costs)
-        self._make_spec(td_params)
+        if generator is None:
+            generator = SVRPGenerator(**generator_params)
+        self.generator = generator
+        self.tech_costs = self.generator.tech_costs
+        self._make_spec(self.generator)
 
-    def _make_spec(self, td_params: TensorDict = None):
+    def _make_spec(self, generator):
         """Make the observation and action specs from the parameters."""
         self.observation_spec = CompositeSpec(
             locs=BoundedTensorSpec(
-                low=self.min_loc,
-                high=self.max_loc,
-                shape=(self.num_loc + 1, 2),
+                low=generator.min_loc,
+                high=generator.max_loc,
+                shape=(generator.num_loc + 1, 2),
                 dtype=torch.float32,
             ),
             current_node=UnboundedDiscreteTensorSpec(
@@ -74,13 +69,13 @@ class SVRPEnv(RL4COEnvBase):
                 dtype=torch.int64,
             ),
             skills=BoundedTensorSpec(
-                low=self.min_skill,
-                high=self.max_skill,
-                shape=(self.num_loc, 1),
+                low=generator.min_skill,
+                high=generator.max_skill,
+                shape=(generator.num_loc, 1),
                 dtype=torch.float32,
             ),
             action_mask=UnboundedDiscreteTensorSpec(
-                shape=(self.num_loc + 1, 1),
+                shape=(generator.num_loc + 1, 1),
                 dtype=torch.bool,
             ),
             shape=(),
@@ -89,50 +84,10 @@ class SVRPEnv(RL4COEnvBase):
             shape=(1,),
             dtype=torch.int64,
             low=0,
-            high=self.num_loc + 1,
+            high=generator.num_loc + 1,
         )
         self.reward_spec = UnboundedContinuousTensorSpec(shape=(1,), dtype=torch.float32)
         self.done_spec = UnboundedDiscreteTensorSpec(shape=(1,), dtype=torch.bool)
-
-    def generate_data(self, batch_size):
-        """Generate data for the basic Skill-VRP. The data consists of the locations of the customers,
-        the skill-levels of the technicians and the required skill-levels of the customers.
-        The data is generated randomly within the given bounds."""
-        # Batch size input check
-        batch_size = [batch_size] if isinstance(batch_size, int) else batch_size
-
-        # Initialize the locations (including the depot which is always the first node)
-        locs_with_depot = (
-            torch.FloatTensor(*batch_size, self.num_loc + 1, 2)
-            .uniform_(self.min_loc, self.max_loc)
-            .to(self.device)
-        )
-
-        # Initialize technicians and sort ascendingly
-        techs, _ = torch.sort(
-            torch.FloatTensor(*batch_size, self.num_tech, 1)
-            .uniform_(self.min_skill, self.max_skill)
-            .to(self.device),
-            dim=-2,
-        )
-
-        # Initialize the skills
-        skills = (
-            torch.FloatTensor(*batch_size, self.num_loc, 1).uniform_(0, 1).to(self.device)
-        )
-        # scale skills
-        skills = torch.max(techs, dim=1, keepdim=True).values * skills
-        td = TensorDict(
-            {
-                "locs": locs_with_depot[..., 1:, :],
-                "depot": locs_with_depot[..., 0, :],
-                "techs": techs,
-                "skills": skills,
-            },
-            batch_size=batch_size,
-            device=self.device,
-        )
-        return td
 
     @staticmethod
     def get_action_mask(td: TensorDict) -> torch.Tensor:
@@ -186,13 +141,7 @@ class SVRPEnv(RL4COEnvBase):
     def _reset(
         self, td: Optional[TensorDict] = None, batch_size: Optional[list] = None
     ) -> TensorDict:
-        if batch_size is None:
-            batch_size = self.batch_size if td is None else td["locs"].shape[0]
-        if td is None or td.is_empty():
-            td = self.generate_data(batch_size=batch_size)
-        batch_size = [batch_size] if isinstance(batch_size, int) else batch_size
-
-        self.to(td.device)
+        device = td.device
 
         # Create reset TensorDict
         td_reset = TensorDict(
@@ -201,15 +150,15 @@ class SVRPEnv(RL4COEnvBase):
                 "techs": td["techs"],
                 "skills": td["skills"],
                 "current_node": torch.zeros(
-                    *batch_size, 1, dtype=torch.long, device=self.device
+                    *batch_size, 1, dtype=torch.long, device=device
                 ),
                 "current_tech": torch.zeros(
-                    *batch_size, 1, dtype=torch.long, device=self.device
+                    *batch_size, 1, dtype=torch.long, device=device
                 ),
                 "visited": torch.zeros(
                     (*batch_size, td["locs"].shape[-2] + 1, 1),
                     dtype=torch.uint8,
-                    device=self.device,
+                    device=device,
                 ),
             },
             batch_size=batch_size,
@@ -217,7 +166,7 @@ class SVRPEnv(RL4COEnvBase):
         td_reset.set("action_mask", self.get_action_mask(td_reset))
         return td_reset
 
-    def get_reward(self, td: TensorDict, actions: TensorDict) -> TensorDict:
+    def _get_reward(self, td: TensorDict, actions: TensorDict) -> TensorDict:
         """Calculated the reward, where the reward is the negative total travel cost of the technicians.
         The travel cost depends on the skill-level of the technician."""
         # Check that the solution is valid
@@ -291,107 +240,3 @@ class SVRPEnv(RL4COEnvBase):
             ).all(), "Skill level not met"
             start = each[1] + 1  # skip the depot
             tech += 1
-
-    @staticmethod
-    def render(
-        td: TensorDict,
-        actions=None,
-        ax=None,
-        **kwargs,
-    ):
-        import matplotlib.pyplot as plt
-        import numpy as np
-
-        from matplotlib import cm, colormaps
-
-        num_routine = (actions == 0).sum().item() + 2
-        base = colormaps["nipy_spectral"]
-        color_list = base(np.linspace(0, 1, num_routine))
-        cmap_name = base.name + str(num_routine)
-        out = base.from_list(cmap_name, color_list, num_routine)
-
-        if ax is None:
-            # Create a plot of the nodes
-            _, ax = plt.subplots()
-
-        td = td.detach().cpu()
-
-        if actions is None:
-            actions = td.get("action", None)
-
-        # if batch_size greater than 0 , we need to select the first batch element
-        if td.batch_size != torch.Size([]):
-            td = td[0]
-            actions = actions[0]
-
-        locs = td["locs"]
-
-        # add the depot at the first action and the end action
-        actions = torch.cat([torch.tensor([0]), actions, torch.tensor([0])])
-
-        # gather locs in order of action if available
-        if actions is None:
-            log.warning("No action in TensorDict, rendering unsorted locs")
-        else:
-            locs = locs
-
-        # Cat the first node to the end to complete the tour
-        x, y = locs[:, 0], locs[:, 1]
-
-        # plot depot
-        ax.scatter(
-            locs[0, 0],
-            locs[0, 1],
-            edgecolors=cm.Set2(2),
-            facecolors="none",
-            s=100,
-            linewidths=2,
-            marker="s",
-            alpha=1,
-        )
-
-        # plot visited nodes
-        ax.scatter(
-            x[1:],
-            y[1:],
-            edgecolors=cm.Set2(0),
-            facecolors="none",
-            s=50,
-            linewidths=2,
-            marker="o",
-            alpha=1,
-        )
-
-        # text depot
-        ax.text(
-            locs[0, 0],
-            locs[0, 1] - 0.025,
-            "Depot",
-            horizontalalignment="center",
-            verticalalignment="top",
-            fontsize=10,
-            color=cm.Set2(2),
-        )
-
-        # plot actions
-        color_idx = 0
-        for action_idx in range(len(actions) - 1):
-            if actions[action_idx] == 0:
-                color_idx += 1
-            from_loc = locs[actions[action_idx]]
-            to_loc = locs[actions[action_idx + 1]]
-            ax.plot(
-                [from_loc[0], to_loc[0]],
-                [from_loc[1], to_loc[1]],
-                color=out(color_idx),
-                lw=1,
-            )
-            ax.annotate(
-                "",
-                xy=(to_loc[0], to_loc[1]),
-                xytext=(from_loc[0], from_loc[1]),
-                arrowprops=dict(arrowstyle="-|>", color=out(color_idx)),
-                size=15,
-                annotation_clip=False,
-            )
-        plt.show()
