+++ rl4co/envs/scheduling/smtwtp/env.py
@@ -13,6 +13,9 @@ from torchrl.data import (
 from rl4co.envs.common.base import RL4COEnvBase
 from rl4co.utils.pylogger import get_pylogger
 
+from .generator import SMTWTPGenerator
+from .render import render
+
 log = get_pylogger(__name__)
 
 
@@ -42,25 +45,15 @@ class SMTWTPEnv(RL4COEnvBase):
 
     def __init__(
         self,
-        num_job: int = 10,
-        min_time_span: float = 0,
-        max_time_span: float = None,  # will be set to num_job/2 by default. In DeepACO, it is set to num_job, which would be too simple
-        min_job_weight: float = 0,
-        max_job_weight: float = 1,
-        min_process_time: float = 0,
-        max_process_time: float = 1,
-        td_params: TensorDict = None,
+        generator: SMTWTPGenerator = None,
+        generator_params: dict = {},
         **kwargs,
     ):
         super().__init__(**kwargs)
-        self.num_job = num_job
-        self.min_time_span = min_time_span
-        self.max_time_span = num_job / 2 if max_time_span is None else max_time_span
-        self.min_job_weight = min_job_weight
-        self.max_job_weight = max_job_weight
-        self.min_process_time = min_process_time
-        self.max_process_time = max_process_time
-        self._make_spec(td_params)
+        if generator is None:
+            generator = SMTWTPGenerator(**generator_params)
+        self.generator = generator
+        self._make_spec(self.generator)
 
     @staticmethod
     def _step(td: TensorDict) -> TensorDict:
@@ -95,15 +88,7 @@ class SMTWTPEnv(RL4COEnvBase):
         return td
 
     def _reset(self, td: Optional[TensorDict] = None, batch_size=None) -> TensorDict:
-        # Initialization
-        if batch_size is None:
-            batch_size = self.batch_size if td is None else td["job_due_time"].shape[:-1]
-        batch_size = [batch_size] if isinstance(batch_size, int) else batch_size
-
-        device = td["job_due_time"].device if td is not None else self.device
-        self.to(device)
-
-        td = self.generate_data(batch_size) if td is None else td
+        device = td.device
 
         init_job_due_time = td["job_due_time"]
         init_job_process_time = td["job_process_time"]
@@ -113,7 +98,7 @@ class SMTWTPEnv(RL4COEnvBase):
         current_job = torch.zeros((*batch_size, 1), dtype=torch.int64, device=device)
         current_time = torch.zeros((*batch_size, 1), dtype=torch.int64, device=device)
         available = torch.ones(
-            (*batch_size, self.num_job + 1), dtype=torch.bool, device=device
+            (*batch_size, self.generator.num_job + 1), dtype=torch.bool, device=device
         )
         available[:, 0] = 0  # mask the starting dummy node
 
@@ -129,24 +114,24 @@ class SMTWTPEnv(RL4COEnvBase):
             batch_size=batch_size,
         )
 
-    def _make_spec(self, td_params: TensorDict = None):
+    def _make_spec(self, generator: SMTWTPGenerator) -> None:
         self.observation_spec = CompositeSpec(
             job_due_time=BoundedTensorSpec(
-                low=self.min_time_span,
-                high=self.max_time_span,
-                shape=(self.num_job + 1,),
+                low=generator.min_time_span,
+                high=generator.max_time_span,
+                shape=(generator.num_job + 1,),
                 dtype=torch.float32,
             ),
             job_weight=BoundedTensorSpec(
-                low=self.min_job_weight,
-                high=self.max_job_weight,
-                shape=(self.num_job + 1,),
+                low=generator.min_job_weight,
+                high=generator.max_job_weight,
+                shape=(generator.num_job + 1,),
                 dtype=torch.float32,
             ),
             job_process_time=BoundedTensorSpec(
-                low=self.min_process_time,
-                high=self.max_process_time,
-                shape=(self.num_job + 1,),
+                low=generator.min_process_time,
+                high=generator.max_process_time,
+                shape=(generator.num_job + 1,),
                 dtype=torch.float32,
             ),
             current_node=UnboundedDiscreteTensorSpec(
@@ -154,7 +139,7 @@ class SMTWTPEnv(RL4COEnvBase):
                 dtype=torch.int64,
             ),
             action_mask=UnboundedDiscreteTensorSpec(
-                shape=(self.num_job + 1,),
+                shape=(generator.num_job + 1,),
                 dtype=torch.bool,
             ),
             current_time=UnboundedContinuousTensorSpec(
@@ -167,12 +152,12 @@ class SMTWTPEnv(RL4COEnvBase):
             shape=(1,),
             dtype=torch.int64,
             low=0,
-            high=self.num_job + 1,
+            high=generator.num_job + 1,
         )
         self.reward_spec = UnboundedContinuousTensorSpec(shape=(1,))
         self.done_spec = UnboundedDiscreteTensorSpec(shape=(1,), dtype=torch.bool)
 
-    def get_reward(self, td, actions) -> TensorDict:
+    def _get_reward(self, td, actions) -> TensorDict:
         job_due_time = td["job_due_time"]
         job_weight = td["job_weight"]
         job_process_time = td["job_process_time"]
@@ -193,39 +178,10 @@ class SMTWTPEnv(RL4COEnvBase):
 
         return -job_weighted_tardiness.sum(-1)
 
-    def generate_data(self, batch_size) -> TensorDict:
-        batch_size = [batch_size] if isinstance(batch_size, int) else batch_size
-        # Sampling according to Ye et al. (2023)
-        job_due_time = (
-            torch.FloatTensor(*batch_size, self.num_job + 1)
-            .uniform_(self.min_time_span, self.max_time_span)
-            .to(self.device)
-        )
-        job_weight = (
-            torch.FloatTensor(*batch_size, self.num_job + 1)
-            .uniform_(self.min_job_weight, self.max_job_weight)
-            .to(self.device)
-        )
-        job_process_time = (
-            torch.FloatTensor(*batch_size, self.num_job + 1)
-            .uniform_(self.min_process_time, self.max_process_time)
-            .to(self.device)
-        )
-
-        # Rollouts begin at dummy node 0, whose features are set to 0
-        job_due_time[:, 0] = 0
-        job_weight[:, 0] = 0
-        job_process_time[:, 0] = 0
-
-        return TensorDict(
-            {
-                "job_due_time": job_due_time,
-                "job_weight": job_weight,
-                "job_process_time": job_process_time,
-            },
-            batch_size=batch_size,
-        )
+    def check_solution_validity(self, td, actions):
+        log.warning("Checking solution validity is not implemented for SMTWTP")
+        pass
 
     @staticmethod
     def render(td, actions=None, ax=None):
-        raise NotImplementedError("TODO: render is not implemented yet")
+        raise render(td, actions, ax)
